Added UI mudole and exported from other modules to UI modules
